Componentise locale dropdown
